refactor: rename app.rb file
